requirements.txt
django==3.1.7
Markdown==3.3.4
django-filter==2.4.0
djangorestframework==3.12.4
psycopg2>=2.7,<3.0
django-ckeditor==5.9.0
Pillow==8.1.2
drf-yasg
requests~=2.25.1